Ok, had some issues with the code and not realizing that I needed to use .reply instead of the queueMsg mechanism.
I might be able to use queueMsg, but it needs some objects that I'm
not sure of. I suspect it would be more polite so I don't flood channel,
but .reply works for now.

Need to see if I can refactor out the @greeter stuff into separate defs,
pretty sure you can, but other errors lead me to this setup of the code